Implemented GET /users/anniversaries endpoint

See merge request foxminded-teamconnect-2/tracker!7

+++ src/main/java/com/ua/teamconnect/tracker/model/dto/PositionDto.java
@@ -5,10 +5,10 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(name = "Position", description = "Data of the position")
 public record PositionDto(
     @Schema(description = "ID of the position", example = "1")
-    Long id,
+    Integer id,
     @Schema(description = "Name of the position", example = "Software Engineer")
     String name,
     @Schema(description = "Department ID of the position", example = "1")
-    Long departmentId
+    Integer departmentId
 ) {
 }
